Changed Wenglor height relative uncertainty values
Decreased Wenglor relative height uncertainty from +/- 1 cm to +/- 1
mm.  Original value was too large.  The relative height uncertainty is
related to the positioning within the Wenglor array, which was
determined to high precision in the lab prior to deployment of the
array.  More of the height uncertainty (which is captured from the
distance sensor) is due to the position of the array relative to the
ground.